MGMT-4443: Adding console-url to AMS subscription.
The end goal is that all cluster's owners will access their clusters via
OCM UI, therefore the console-url needs to be updated in the
subscription as soon as the service gets it.

For this commit I also needed to update ocm-sdk package's version
because the old version was missing some fields in the Subscription
object.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>