MGMT-4623 Publish index image for assisted service operator bundle (#1348)
* MGMT-4623 Publish index image for assisted service operator bundle

to quay.io

The index is published to https://quay.io/repository/ocpmetal/assisted-service-index

Publishing the index enables users to use the latest published operator
bundle in quay.io.

Signed-off-by: Richard Su <rwsu@redhat.com>

* MGMT-4623 Add BUNDLE_IMAGE and INDEX_IMAGE to skipper.env

* MGMT-4623 Move index build and publish to after the bundle is published

When building the index, the opm tooling pulls the bundle image.
If the bundle image has not been published, the index build fails.