MGMT-4697 The DownloadClusterKubeconfig API switches kubeconfig-noingress and kubeconfig
Up until now, the client (i.e. UI) had to manage the logic for when the
kubeconfig file is available (when the cluster is installed). In case
not - the client had to download kubeconfig-noingress directly.

Instead, DownloadClusterKubeconfig API manages that logic.
In case cluster's kubeconfig file is available then it will download it,
otherwise it will try to download the kubeconfig-noingress file.