NO-ISSUE Remove minikube code and PROFILE variable (#1336)
Our Python scripts were dependent on minikube tool.
Minikube requires a host with a KVM.

Making our scripts work regardless of the cluster.
A simple kubectl command approach that is configured to some cluster.
That cluster could be local or remote.

Some lightweight approaches like k3d and kind deploy a cluster on top of dockers,
would be able to work as well.