MGMT-4850 set InstallationDiskID based on RootDeviceHints (#1461)
This PR adds the ability to select an installation disk based on the RootDeviceHints provided by
the BMH. Note that only the hints supported by OpenShift (BareMetal Operator, specifically) were
implemented here. The rest of the filters and operations supported by ironic are not needed for
now as they are not even used today in the OpenShift case.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>