MGMT-4921 Rename InstallEnv --> InfraEnv
As the assisted-service KubeAPI approaches a point in which it starts to
integrate with other projects, we make final changes to the API naming.

This started with https://github.com/openshift/assisted-service/pull/1363
to rename group and version.

This PR renames InstallEnv to InfraEnv.
Rename will include both CRD and its matching controller.