MGMT-4141: cluster controller: clear any APIErrors after successful update
When attempting to patch a ClusterDeployment, UpdateClusterInternal is called.
If fails due to bad configuration, an APIError will appear in the status condition.
Currently, there is no place that clears that error after the configuration is
fixed and cluster update completed successfully.
We don't want the user to be confused because he did not know which errors are relevant
and which aren't.
To this PR I added this error clearing, which will happen only if UpdateClusterInternal
completed with no errors.