MGMT-4740 Encrypt traffic to assisted service
This commit adds an annotation to the service which will cause a
tls secret to be created in the same namespace valid for that service
name. This secret is used to configure assisted-service to start an
https server.

Additionally, the route is configured to use "re-encrypt" tls
termination which will use the router's certificate to communicate
with the client, then re-encrypt the request from the router to the
service using the service cert.

Docs on the service annotation:
https://docs.openshift.com/container-platform/4.7/security/certificates/service-serving-certificate.html#add-service-certificate_service-serving-certificate
https://issues.redhat.com/browse/MGMT-4740