OCPBUGSM-26430 - fix dns resolution for bootstrap node configured with bond interface
Bonding interface may take a long time to be up, therefore NM will set its transient name
to localhost, since no default link is available. Afterwards, API VIP is set on that interface which causes NM to check if
there is an update to hostame ( and there will be since link is up and it can perform reverse lookup. In case hostname has changed,
NM tries to recalculate the resolv.conf file contents, and it will set it to use the DNS ip provied by configuration( external), and will remove
the local ip 127.0.0.1, set by the prepender script. Since the IP change is external, and the resulting hostname is not valid for hostnamed service,
the prepender script is not run.
Solutuon: set the NM configurion hostname mode to none, which will block Nm from trying to resolve/chec hostname updates. In addition, this configuration
will not be propagated to the RHCOS upon node reboot