MGMT-5176 - Fix support for disabling hyperthreading (#1500)
hypethreading configuration is set in the install-config and is used
by the installer code to produce ignition kernel param for RHCOS.
by default hypethreading is enabled by the RCHOS, so installer code needs
to configure kernel arg only in case of hyperthreading disabling. the valid values
for hyperthreading parameter in ignition is Enabled and Disabled. In case the value is missing,
it is Enabled by default