MGMT-5240 - Add recently released 4.8.0-fc.0 to the default versions JSON (#1518)
This will make it so that 4.8.0-fc.0 is available in staging/production
environments as a beta, non-default version.

Also updated the `hack/get_ocp_versions_for_testing.sh` script so it no
longer appends the "fake" 4.8 image for single-node.

However the override code is still kept around behind this gate to provide
an example of how we can do a similar override in the future.