NO-ISSUE: Add APIGroup and resource to kube_api leader election role
The coordination.k8s.io apiGroup and the leases resource were missing from the kube_api role.

The reason why this worked until now is still unknown to me but I'm guessing it has to do with the use of
older versions of the depenencies being updated in this PR.

Ideally, we should not duplicate the role definition, which exists under config/rbac as well. We'll fix this
in a follow-up commit.

Signed-off-by: Flavio Percoco <flavio@redhat.com>