MGMT-4741 Use the ingress cert to ensure trusted https connections
According to
https://docs.openshift.com/container-platform/4.7/security/certificate_types_descriptions/ingress-certificates.html#workflow
this configmap is the authoritative source of the certificate being used
for ssl on routes.

This commit also configures the SERVICE_CA_CERT_PATH environment
variable on the assisted service to point to the file in the new
configmap. This allows us to enforce trusted connections to the service
from the agent and controller.