NO-ISSUE Set `online: true` and disable autoamted cleaning in the BMH resource (#1512)
* NO-ISSUE: Bump baremetal-operator and hive dependencies' versions

This commit updates both the Hive api and the BMO dependencies to bring in changes from
the last rebase. Namely, the ability to disable cleaning on the BMO.

As a consequence of the BMO update, the commit also updated the controller-runtime version to
0.7.2, which introduced several changes to its APIs.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* NO-ISSUE: Disable automated cleaning and set online true in BMH

This commit makes sure the BMH automated cleaning mode is disabled and
that the Online parameter
is set to true. The latter makes sure the node is booted once the ISO
URL is set.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* NO-ISSUE: Add APIGroup and resource to kube_api leader election role

The coordination.k8s.io apiGroup and the leases resource were missing from the kube_api role.

The reason why this worked until now is still unknown to me but I'm guessing it has to do with the use of
older versions of the depenencies being updated in this PR.

Ideally, we should not duplicate the role definition, which exists under config/rbac as well. We'll fix this
in a follow-up commit.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>