MGMT-4273 Create secret key to enable local AUTH_TYPE in operator (#1467)
* MGMT-4273 Create secret key to enable local AUTH_TYPE in operator

This enables authentication between assisted-service and agents.

Creates a secret called "agent-local-auth-key" containing
ec-public-key.pem and ec-private-key.pem as data.

In the operator, sets AUTH_TYPE=local in the assisted-service
deployment.

* MGMT-4273 Add agentserviceconfig_controller_test

* MGMT-4273 Remove recorder and add comment on why route.Spec is changed

The recorder is removed because we don't actually have events that
need tracking. And because a working recorder cannot be created in the
tests.

var recorderProvider recorder.Provider
recorderProvider.GetEventRecorderFor("agentserviceconfig-controller"),

results in "runtime error: invalid memory address or nil pointer dereference"

Added comments explaining why route.Spec = routeSpec cannot be used
in newAgentRoute's mutateFn.

* MGMT-4273 Refactor test to focus on ensure instead of reconcile

This makes it easier to construct additional tests on ensure
instead of funneling resource tests through reconcile.

* MGMT-4273 Add recorder back to the controller. Remove unnecessary

object cleanup.

Was supposed to just remove the recorder from the test and from
recording the reconcile success event.

Remove object cleanup from tests. The cleanup is not required
because the objects are stored in memory by fake client and a
new client is created for each test.

* MGMT-4273 Use SetStatusConditionNoHeartbeat and remove route change

SetStatusCondition was changed to SetStatusConditionNoHeartbeat.
Missed it in the rebase.

Remove route change because testing reconcile was removed from
this PR.