MGMT-4740 Encrypt traffic to assisted service (#1506)
* MGMT-4740 Encrypt traffic to assisted service

This commit adds an annotation to the service which will cause a
tls secret to be created in the same namespace valid for that service
name. This secret is used to configure assisted-service to start an
https server.

Additionally, the route is configured to use "re-encrypt" tls
termination which will use the router's certificate to communicate
with the client, then re-encrypt the request from the router to the
service using the service cert.

Docs on the service annotation:
https://docs.openshift.com/container-platform/4.7/security/certificates/service-serving-certificate.html#add-service-certificate_service-serving-certificate
https://issues.redhat.com/browse/MGMT-4740

* MGMT-4740 Use https for service health checks

A TCP check on the https server caused an error message every 10
seconds:
2021/04/20 20:27:39 http: TLS handshake error from 10.217.0.1:53612: EOF
2021/04/20 20:27:49 http: TLS handshake error from 10.217.0.1:53878: EOF

Additionally, this update makes the liveness and readiness probes match
the ones we have in the cloud deployment.