MGMT-4321 Prevent mixing IP families
Although OCP supports dual stack, more work is needed to properly
support it in Assisted Installer. For now we will block combinations
of IPv4 with IPv6 addresses in the API, and make sure we do not
automatically pick up an IP address or CIDR that do not match the
selected address family.