MGMT-5359 High CPU usage of Assisted-Service with 500 clusters and not hosts (#1573)
The change handles the flow when there are many clusters that do not need to be updated during RefreshStatus.
This flow should be fast as possible.  Testing showed that many db queries cause high CPU usage and latency in cluster monitoring.
So during the above flow the aim for clusters that do not need to change - to query the database much less than
the number of clusters.  Currently we use 100 clusters batch.
The solution is in cluster monitoring to query one per batch for all cluster operations in the loop (connectivity-groups, refresh-status).
This data should be used without additional database query unless update is attempted.
The main idea for the change is that clusters are queried only once in the monitoring loop with all the details needed (hosts,
operators etc.). These clusters are used for all the monitoring flow in case that no update is attempted. Basically it contains the following changes:

- The refresh status cluster uses cluster from the initial query.
- In case cluster is updated during refresh status, the result of the update is taken from the update function.
- The loadCluster was removed when preparing cluster. The cluster is taken from the initial query.
- The clusters for connectivity check calculation are taken from the initial query.