NO-ISSUE Skip agents with an empty MacAddress
It is possible create Agents that have an empty MacAddress. If this happens, and there are BMH's that
don't have a BootMACAddress set, then the agent will be associated with that BMH and hardwaredetails will
be set on the resource.

This commit skips Agents that have an empty MacAddress just for extra safety meassures

Signed-off-by: Flavio Percoco <flavio@redhat.com>