MGMT-3512 Skip garbage collection jobs in operator deployment (#1582)
In operator deployment, when cluster entity is removed from the
database by the garbage collector, the clusterdeployment controller
will attempt to reconcile the clusterdeployment CR by re-creating a new
cluster entity in the database for it.
Therefore, the garbage collection job is not useful in operator
deployment and needs to be skipped.

Signed-off-by: Moti Asayag <masayag@redhat.com>