MGMT-5386: Fixing OCM-Operations Grafana dashboard. (#1583)
I was querying for all metrics in bucket with le!="+Inf" but the
histogram is accumulative, means that I was counting the same metrics
over and over.

I just removed the 'bucket' suffix and used 'count' suffix instead,
after all we want to count also the "failing" requests (with le="+Inf").

Signed-off-by: Yoni Bettan <ybettan@redhat.com>