OCPBUGSM-28006 Fix vxlan vmware incorrect checksum in assisted installer (#1580)
Add manifest to create a bash script file /etc/NetworkManager/dispatcher.d/05-disable-tunnel-offload
The manifest creation is invoked for clusters with vmware hosts only
The script checks if the driver is vmxnet3 and that the settings tx-udp_tnl-csum-segmentation and tx-udp_tnl-segmentation
can be updated, If these settings can be updated, the script sets the to off.
The script is relevant (and checks it) for "up" action only.