MGMT-5174 Subsystem tests for epic MGMT-3563 (#1526)
Tests that are validating the effect of successful or failed disk-speed-check response
on the flow of the installation. In case of failed response, the host should
return to insufficint as well as the cluster.  In case the all hosts succeeded preparing (successful disk-speed-check response),
installation should continue. The successful flow is already tested as part of the other installation tests.