NO-ISSUE: Adding debug info for failing validations subsystem tests.
GinkgoWriter is an io.Writer that emit its content to the console if a
specific test fails.

In our case, we have some flaky tests due to the fact that multiple
metrics are filtered instead of a single one as part of the test. This
new info will help us understand why more than a single metric was
filtered.

Signed-off-by: Yoni Bettan <ybettan@redhat.com>