MGMT-3700 Add dedicated subsystem tests for BMAC day0  (#1584)
* MGMT-3700 Update the BareMetalHost CRD

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* NO-ISSUE add more debug messages for BMAC

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* NO-ISSUE Skip agents with an empty MacAddress

It is possible create Agents that have an empty MacAddress. If this happens, and there are BMH's that
don't have a BootMACAddress set, then the agent will be associated with that BMH and hardwaredetails will
be set on the resource.

This commit skips Agents that have an empty MacAddress just for extra safety meassures

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* MGMT-3700 Add dedicated subsystem tests for BMAC day0

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>