MGMT-4261 Deregister Host on agent CR deletion
This change implements a finalizer in the agent controller.
Pre-deletion logic includes host deregistration from the backend.

The controller will also place a deletion Finalizer when:
1. clusterDeployment cannot be retrieved for the reconciled agent.
2. Backend cluster cannot be retrieved for the reconciled agent.
3. The agent (host) is no longer a part of the  backend cluster (day2 scenario)

Lastly, to fetch a host when we have no information about its cluster,
I added a method to query it by host ID.
Once we implement MGMT-6006 we may change it to kubeKey.