MGMT-5331 Subsystem test for filesystem usage metrics
when the configured storage is 'filesystem', a filesystem usage metrics
are reported when filesystem is modified.
The test will verify that the expected metric is collected after ISO is
generated (which writes file to disk and therefore changes the
filesystem usage) and its value is of type float and in the expected range.

The test should be skipped for non-filesystem storage, since no metrics
should be collected in that case.

Signed-off-by: Moti Asayag <masayag@redhat.com>