OCPBUGSM-26546 Validate length of base DNS domain
Validate the length of base DNS doman and its labels.

The lenght must adhere to the 255-bytes limitation while
allowing also for auto-generated DNS names that an OCP cluster
needs, which are created by concatenating a reserved application name,
a cluster name, and a base domain.

Example:
'alertmanager-main-openshift-monitoring.apps.test-cluster.example.com'
where 'test-cluster' is the cluster name and 'example.com' is the base
domain.