OCPBUGSM-28683: Allow to signal when the Reconcile should return
This adds a Complete function to the reconcileResult interface so that it will be
possible to signal the main Reconcile function to return early after some of the
steps. This will help reducing queries, reconcile steps, and checks when we already
know that we are not ready to run those.

Signed-off-by: Flavio Percoco <flavio@redhat.com>