OCPBUGSM-26820 Add database connection retry (#1673)
In operator deployments the assisted-service and postgres
are deployed in the same pod in separate containers. Currently,
assisted-service does not wait for the postgres to be ready
before attempting to open a connection to postgres, and often
times the first start fails causing the container to restart.

Because both are in the same pod, initContainers cannot be used
to wait for postgres to be ready before starting up
assisted-service.

Database connection retry is added to give more time for postgres
to be ready. Instead of failing the container immediately (by
logging a fatal error) when the db connection cannot be established,
it will now retry every 2 seconds for up to 10 seconds. After 10
seconds, if a connection has not been established, it will log a
fatal error.

A 10 second timeout is chosen to be less than the time (30 seconds)
when the LivenessProbe on the assisted-service container kicks in.