MGMT-6067 Add a cache for the local authenticator (#1683)
We will now only go to the database to check for presence of the cluster
from the token claim once every 10 minutes.

We don't store any data in the cache because the presence of the key is
enough to authenticate the request. We also don't cache failed responses
as we don't want it to take 10 minutes for a request to succeed if a
cluster is somehow not present, but will be in the future.