MGMT-4966 Map the /run/media/iso directory to the next_step agent component (#1685)
When a media disconnection error occurred the system became unstable and the step execution might be stuck on IO errors or failed with an unfriendly message.
Media disconnection occurred when a network drive (mostly PXE) has any errors (Network issue) and the controller disconnects it automatically.
Since this error is common to all the steps, we add this verification to the agent.
/run/media/iso is the ISO directory which mounted from the CDROM drive