Bug 1957227: Allow overriding defaults via provided ConfigMap
This PR fixes a bug where a custom ConfigMap configured via the
annotation
`unsupported.agent-install.openshift.io/assisted-service-configmap`
allowed to define additional variables, but did not allow to override
the ones provided by default by the service. This was causing a
confusion in scenario when custom ConfigMap contained e.g.

```
SERVE_HTTPS: 'False'
```

but the real state of the deployment was not reflecting it correctly

```
oc rsh deployment/assisted-service
env | grep SERVE_HTTPS

SERVE_HTTPS=True
```

With this PR we are creating a ConfigMap with all the service defaults.
When creating a service container, EnvFrom field will be populated with
both default ConfigMap as well as the one created manually (if present).
In order to simplify the logic, service secrets are not mounted via
EnvFrom but via Env.

This change gives customer an ability to override an arbitrary value
without taking ownership of the whole configuration.

The initial issue was caused by the order of applying variables by
Kubernetes, i.e. Env always taking precedence over EnvFrom.

Closes: OCPBUGSM-28781