MGMT-6282: Split openshift_cluster_id and status 'Active' update in AMS.
openshift_cluster_id is know at an earlier stage, before installation
begin, and this is when we need to update it in AMS subscription.

Currently we are experiencing a race with the subscription created by AMS
when cluster telemetry arrives.

In case the installation fails, we will remain with an obsolete
'external_cluster_id' (aka openshift_cluster_id) in the subscription
until the user restarts the installation (and the openshift_cluster_id is
overwritten) or until the cluster is deregistered from AI DB (and
therefore the subscription deleted in AMS).

Signed-off-by: Yoni Bettan <ybettan@redhat.com>