NO-ISSUE Combine job implementations
The only real difference between the local and regular job
implementations is an assumption that we're using the local s3 client.
Before this change that was incredibly confusing because in the
implementation of the jobs used the s3 client as an interface rather
than a particular implementaton.

This commit leans into the interface model and removes the reliance on
the generator working directory needing to be the same as the local s3
client base directly. This is done by copying the files after they are
generated using generator.UploadToS3.