NO-ISSUE Add Elastic APM agent (#1707)
* NO-ISSUE Add Elastic APM agent

This is the first commit in a series of commits tha twill add
support for Elastic's APM agent throught assisted-service.

Elastic APM will allow for tracing performance metrics for assisted-service at
every level of the application. To start with, we are focusing on the following:

- HTTP Tracing
- Logs tracing
- DB Tracing
- Specific functions tracing

This commit adds the http and logs tracing. Both of them are behind a feature flag
called EnableElasticAPM as a way to prevent it from tracing requests in every environment. The
target is to enable it for Red Hat's hosted deployments only.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* NO-ISSUE Use the operationID instead of the route pattern

This will give us human readable names for the APM metrics and dashboards. Basically,
ListClusters instead of

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>