NO-ISSUE Remove assisted-service finalizer as part of destroy script
Deletion of a localvolume could get stuck if its finalizers don't
finish.
By default, localvolume has a finalizer that protects the data and leads
to a graceful shutdown. While this graceful shutdown might be stuck and
the destroy script as well.
Since this script is deleting everything, we are not afraid to lose
data, hence the forceful patch of removing the finalizers.