NO-ISSUE Remove kube job references and combine job types (#1737)
* NO-ISSUE Remove unused kube job handling

The actual job creation was removed in e805401970e98676bd145343952dea949ca5b4a3
this is just all the extra stuff that we don't need to keep around.

* NO-ISSUE Promote workdir to main options

This is being used for both filesystem storage and the local job
working directory. It's a bit strange that we have to reference the job
config when creating the storage client.

* NO-ISSUE Combine job implementations

The only real difference between the local and regular job
implementations is an assumption that we're using the local s3 client.
Before this change that was incredibly confusing because in the
implementation of the jobs used the s3 client as an interface rather
than a particular implementaton.

This commit leans into the interface model and removes the reliance on
the generator working directory needing to be the same as the local s3
client base directly. This is done by copying the files after they are
generated using generator.UploadToS3.

* NO-ISSUE Move the job implementation to generator

Not sure why the interface was in a separate package