OCPBUGSM-28663 Create an iso download endpoint with .iso at the end - bz-1956421 (#1744)
* OCPBUGSM-28663 Create an iso download endpoint with .iso at the end

This adds a new endpoint with the same implementation as the existing
GET/HEAD /clusters/{cluster_id}/downloads/image but suffixed with .iso to
comply with some BMC requirements.

https://bugzilla.redhat.com/show_bug.cgi?id=1956421

* OCPBUGSM-28663 Use the new endpoint for the download url

This sets the cluster download url to the path ending with `/image.iso`
rather than the old `/image` endpoint.

The BMH controller gets the url from this same location so by changing
it here we're also changing it there.

https://bugzilla.redhat.com/show_bug.cgi?id=1956421