OCPBUGSM-29303: fix crash on undefined pull-secret (#1755)
Pull secret is optional in cluster deployment but it is required for assisted service.
Get pull secret will get a reference instaed of the name directly and will check if refence is not nil
if it's nil it will return input error that will cause sync error in the sync condtion.

Added new error type to the controllers inputError type that is changed
as internal error with `IsHTTP4XXError`.

Rename `IsHTTP4XXError` to `IsUserError`