Bug 1962957 - Check the response status code when downloading files
Before this change if the GET request here failed in a way that didn't
produce an error we would continue on to try to extract the boot files
from the error response body.

According to https://golang.org/pkg/net/http/#Get non-200 responses do
not generate an error so we need to check this ourselves.

https://bugzilla.redhat.com/show_bug.cgi?id=1962957