MGMT-4668: Generate OLM opeartor manifest to bootstrap ignition (#1609)
This PR add a possibility to create manifests from assisted-installer.
In some cases it may be useful to create manifest from
assisted-isntaller instead of openshift-installer. For example when
installing CNV/LSO/OCS we need to have predefined CRDs for the
operators, because at the time of the operator is being created the CRDs
are not yet created by the OLM.

This PR add support to read the files from custom_manifest.yaml
and fetch the content of those files
and create them via dynamic client. Since assisted-installer-controller
runs with limited permissions, every resource which should be created by
assisted_installer_controller must be added to the clusterrole.

Signed-off-by: Ondra Machacek <omachace@redhat.com>