OCPBUGSM-29066 Remove cluster resources after clusterdeployment deletion
When cluster is deregistered, either by clusterdeployment or by the API,
the cluster record is softly deleted, meaning the  attribute 'deleted_at'
in the DB is being set.
It is the responsiblity of the garbage collector to run periodically and
remove all of the cluster's resources.
For that purpose, the garbage collection should be enabled for clearing
cluster resources in operator deployment as well.

Signed-off-by: Moti Asayag <masayag@redhat.com>