OCPBUGSM-29476 Trigger BMH's deprovision when the ISODownloadURL changes
There are cases when the InfraEnv or ClusterDeployment parameters are chagned, which requires AI
to rebuild the discovery ISO. When this happens, BMAC must make sure that the BMH is properly updated
and that care is taken so that the existing image can be properly deprovisioned before the new one is
added.

This PR does that by comparing the existing ISODownloadURL value with the one set in the BMH. When they difer,
BMAC assumes that a deprovision of the existing URL must happen and the new one should be, eventually, provisioned.

Note that the AI base url doesn't change and this PR relies on the signature of the URL to change. This will only
happen when AI is running with authentication enabled. In a non-auth environment, this workflow won't work.

Signed-off-by: Flavio Percoco <flavio@redhat.com>