MGMT-6605 Create disks with WWN
The way disks were created didn't guarentee the disks locations as we
expect them to be for the localvolume.

In order to make an absolute expected destination, we should provide a
WWN hash that would be available as a link on
/dev/disk/by-id/wwn-<hash>.

The WWN token has to begin with a number, so it was prefixed with an
hardcoded 0 and afterwards 15 bytes of a digested hash from the original
disk name.