MGMT-6633 set db connection pool (#1810)
Right now we are using unlimited connections with default number of allowed idle. This default value is 2. When we have burst or more than 2 routines that reaches DB in parallel assisted-service will create new db connection and then close it after finishing db call. Creating this DB connection when we have some burst of apis is relatively expensive and can increase api call for a second.  
Max open connections number will be set to 100 by default (it is default value in postgress) but in cloud default will be 600 as we have db.r5.large instance with 16G ram and by https://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/CHAP_Limits.html#RDS_Limits.MaxConnections maximum number of open connections should be less than 1800 and we have 3 instances so it should be 1800/3