OCPBUGSM-28683: Allow to signal when the Reconcile should return (#1678)
* NO-ISSUE: Set a name for the bmac controller

This is to have a more descriptive name in the logs. Even if BMAC logs are separate
from the BMO ones, it's not nice to have both controllers using the same name.

https://github.com/openshift/baremetal-operator/blob/d22c5f710cdce4a30f3d8a8457d45ad53dd3f3a2/controllers/metal3.io/baremetalhost_controller.go\#L1321-L1329
Signed-off-by: Flavio Percoco <flavio@redhat.com>

* OCPBUGSM-28683: Allow to signal when the Reconcile should return

This adds a Complete function to the reconcileResult interface so that it will be
possible to signal the main Reconcile function to return early after some of the
steps. This will help reducing queries, reconcile steps, and checks when we already
know that we are not ready to run those.

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* OCPBUGSM-28683: Add image url to subsystem test to skip BMH reconcile

Signed-off-by: Flavio Percoco <flavio@redhat.com>

* OCPBUGSM-28683: Address review comments

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>