OCPBUGSM-29623: Configure IPv6 DHCP in dual-stack
In order for NICs to retain IPv6 addresses received from DHCP
across reboots, `ipv6.dhcp-duid=ll` must be configured. We did
it for IPv6-only hosts, but it should be done also for dual-stack
hosts. For dual-stack OCP installation it only matters if all the
hosts of a cluster are dual stack.