OCPBUGSM-29721 Keep Image URL for existing image with auth=local + S3 (#1861)
The InfraEnv controller passes ImageCreateParams to the backend per
each reconcile, letting it decide whether or not it should generate a new image.

In turn, the backend clears up the image download URL, which eventually filled
the same URL when auth=none. However, when Assisted-Service configuration is
set to with auth=local and s3, a new URL will be generated both with a new or re-used image.

The above scenario caused the BMH, which is watching infaEnv and specifically
looks at the URL to retrigger hosts install.

This change will not clear up the URL only if the image does
not exist and will generate a URL only in a case of a newly generated image.

Lastly, the change included ed some debug-level
prints that helped to understand the issue.

Co-authored-by: Michael Filanov <mfilanov@redhat.com>

Co-authored-by: Michael Filanov <mfilanov@redhat.com>