MGMT-6607: Fix OCS Cluster Validation and remove minimal mode
The ValidateCluster should not check for resource requirements. It should validate that there are at least 3 disks for OCS, the number of hosts with OCS disks >=3 and that the number of disks are a multiple of 3.
Minimal mode needs to be removed are ValidateCluster does not take into account requirements of other operators, so deciding on minimal or standard mode based on resources is not possible.

Signed-off-by: Rewant Soni <resoni@redhat.com>