NO-ISSUE Add a PULL_REQUEST_TEMPLATE for assisted-service (#1842)
This template will be proposed every time a new PR is created. The
author of the PR will be prompted with it and expected to fill in the
suggested sections, as needed.

None of these sections are required but they are definitely recommended
as they will make the review process easier for everyone.

More on PR templates here: https://docs.github.com/en/communities/using-templates-to-encourage-useful-issues-and-pull-requests

Signed-off-by: Flavio Percoco <flavio@redhat.com>

Co-authored-by: Flavio Percoco <flavio@redhat.com>