Bug 1966546: KubeAPI keep day1 cluster after install
Currently, once the cluster is installed, the day1 backend cluster is deleted.
Therefore the events and logs are not available anymore.

In this patch, the day1 is not deleted.
Note that day2 functionality will not work anymore and will need to be
revisited when the support for it will be required.
Signed-off-by: Fred Rolland <frolland@redhat.com>